Added in the sum all function to sum all numbers within a slice and put the result in an array.
Learning how to properly use the make command. Looks like go allow for the creating of a new empty data structure using make and then supplying the size of the created slice.

This could be useful if you wanted to make a copy ahead of time as things will be allocated ahead of time for the length of the previous structure.