Go Tests
Created a wallet and added tests. One of the key focus points here is that go by defnintion passes each value by **copy**. Meaning that we are modifying a value that was passed in via the test but not by its own internal state. We have to leaverage **pointers** to make sure that we are maintaining the state within the class when we pass things around. Now we are pointing to the same value (balance) and we can make our update accondingly.

Unlike C and C++ go will cast to the proper systax for us. So when we call wallet.balance on the main type it knows to use the pointer notation instead of the object reference. This allows us to keep our code clean and be real intentional for how we work with pointers.