Go Test
Added in the tests for a dictionary to create a search. Really liking the type definintion in golang.
I overrode the map[string]string type to be a Dictionary instead. This means I can use it as a "dictionary" but it has all of the code of a map. Pretty cool