Go Test
Added in the tests for a dictionary to create a search. Really liking the type definintion in golang.
I overrode the map[string]string type to be a Dictionary instead. This means I can use it as a "dictionary" but it has all of the code of a map. Pretty cool.
Also you will notice that there is no pointer reference here. I'm not worried about that right now because I am not mutating the data. This will be a full copy of the map (slow i know) but for this example it is not important. The tutorial may have me working with the pointer soon.