Go Training
Ah another gotcha... sorta. So what I'm forgetting here is if I want to work with an io.Writer I need not to work with it directly. io.Writer only implements the Write interface. That means that I am passing the interface which cannot be a pointer. I keep trying to pass a pointer but as you can see in the test that is useless. What I am really doing is getting the addess of an interface (bytes.Buffer in this case)  and passing that to the function. This will then go and be used in the test. In the main code I'm passing os.Stdout instead. This will use the os.Stdout interface (which implements Write()) and use that instead.

fmt.Fprintf will just print out a formated string. Unlike f.Printf() it does not default to any specific writer interface. Which is cool. So now I'm passing in the os.Stdout inteface and making things happy all over again.

Key Point to Remember: io.Writer is an **interface**. You supply the **interface's address** to the function (which can happen implicitly in some cases) and then this can be leveraged by the string formatter library fmt to get you where you need to go.