Go Training
Big win for TDD here. I need to get the end to end and the plumbing working first. Nice thing about this is that I have the plumbing working in the last test, I was able to get a value from the function. This means that I can literally call that out to say "well the function works, lets iterate on it". So for this example I decided to just call the 3 - go in a set of statements that will allow me to do a countdown. Which is cool and not at all Dry. Which is fine for now. I get the string without new lines as I can add those later and really just get my test string working. This is bliss.