Go Train Mocking
So GO does have template syntax with the backtick operator! So much easier if you ask me (although I kind of already knew that). I was able to get it to print like it would from the console even though it messes up the formatting. That's a sad thing, but sometimes you have to break the rules. I wonder what gofmt would do...