Go Training Mocking
So this time we created an array that allowed us to keep track fo the calls in the order that they were called. I accidently deleted the first test and created a sub test. I will add that back as it will be good practice for writing go code. Utilizing reflect.DeepEqual() has also helped to simplify the array comparison so that we don't have to worry about comparing everything in place.

Additionally I created a new struct that implemented both the sleep and write calls. This mega spy now can keep track of everything (if need be) or do one thing on its own if needed. I its a lot of extra code to do this kind of work. I hope that we will get into a go mocking library, but I can't be certain one exists.