Go Training Mocking
Added back in the first test from memory. I like the expressive syntax of go and its interface ability. I need to learn the terminology to be more effective in talking about go as well. I believe spy := &CountdownOperationsSpy{} is called a struct literal. It is staying "store the address of this struct in this variable". The {} really is just an empty struct so all of the values will be initialized to its logical empty. In this cases s.Calls will just return an [].