Go Train Mocking
I should listen to what the prompts are telling me instead of doing my own thing. Here we crated a new type and then we tested the new type to makek sure that it satified the sleep interface and the mock did the same thing. So basically we spied on our newly created type. Then we modified the behavior of the main program to add in that new configuration **leaving the calling function alone**. I didn't do that and jumped right into implementation (again!).

[Improve]
Need to focus on writing the tests first and not the code. Its more of muscle memory that needs to be created here instead of going headlong into the code. This is a tough habit to break and I'm not suprised that I do it. But now that I 've called it out I can be more aware of it.