Go Train Concurency
Added in the example code from the concurrency section. I completely forgot about Benchmark as a way to write tests in go! Which makes sense. The cool thing about that is that I can name a test wtih the Benchmark keyword and find if things are slowing down.

Running a benchmark is as simple as runnin go test -bench=.
What this does is find all of the files *_test.go and if they have the Benchmark keyword in the name they will go and run the benchmark code.