Go Train Concurrency
Go is a facinating language. As I knew before you can easily use the keyword ``go`` in front of a function to make it parallel. You can then use a anon function to make it into a parallel routine. In fact this anon function is handy because it maintains the lexical scope of the function, so anything declared in the function body above will be perserved. This is why we are able to not create results as a function parameter to the anon function but rather just call it like it was already part of the function body. Pretty cool.