Go Train Concurrency
Interesting again. What is really interesting here is that the lexical scope is perseved... but not in the way that we expect. In fact what the older version of code did was pass the same url string multiple times to the function. What we need to do is pass the url to the invocation string of the function and then add that to the function signature. Just like in Javascript with the IIFC we need to pass a value to each go thread to make it work predictibily. By doing this we have the ability to make each goroutine have its own copy of the url instead of sharing the url like in the past function.