Go Train Concurrency
So in order to combat a certain race condition we need to have the ability to pass data to a channel. A channel is created using the make() keyword as we want to create a new non null memory address to do our work in. It really is just a lightweight message container. We created an ADT to hold our data and then we passed in our result in the same lexical scope as our function. This channel then waited until a result was creatd for each individual url passed into the function and sent that data along the channel. A channel is just a working area for results of a function call at this point. Then we read from the channel for each individual write and store the results in the results map and return that map to the caller.