Go Train Select
Starting this off right now to see if we can get a race condition going between two urls. I assume from the chapter title we will be focusing on using select fort this. What I really like about this book is that you use the library in real ways. I calculated duration with time.Now() and time.Since(now). Pretty neat and clean if I ever need to track the time of anything.