Go Train Select
Okay so on last commit we learned we can get a time.Duration from time.Now() being fed into time.Since(now). Which is cool. This time we are seeing that we can mock out a test http server in go with the "net/http/httptest" package. There is some studder there but this was an early package. What it allows us to do is create a server that will then create a mock handler function that we can leverage instead of using a real server. In this case we have two servers (probably could be a factory here) and each server either waits or does not wait to return a response (aka 200 OK). Then we test the response and get our answer back.