Go Train Select
Making this commit seperate for a purpose. The defer statement is called at the END of a containing function. So we want to make sure that we are doing all of our work but we shut the server down before it the test is finsihed. The only way to do that effectively is to use the defer statement. This means that we can keep the code coupled with definition but the function will only be called at the end. Pretty niffty. Basically setting up a function "TO DO" list as we write our function