Go Train Select
Very interesting. So select in go can work like a normal switch but also allows for reading a set of pipes. Abstractly we have set up a set of pipes that is listening for something to write to the pipe. We send the data in the select to the method and the method returns to us a channel. This gets held hostage in the select statement as a sort of list. Then essthe select is listening for behavior on the channel we created. This means that if something were to happen (a close in this case) we can detect it on the registered channel of the select and take action on it. Wow.