Go Train Select
Added in a case where if the requests take longer than 10 seconds that we time out and send an error. This again is something we can do with the select statement. time.After must return a channel much like a timeout in javascript. So we care able to add it to our listenting channel list. If the time is exceeded it will close the channel and allow us to continue on with the error message which makes the code quite clean.

What we have now is a problem in the main code where we sleep for too long but our test is not seeding that value. This could lead us to having to wait a total of 12 seconds for a test to pass... not good.