Go Train Select
So this was the tail end of the problem. I've never really thought of this technique before Friday, but it makes sense. I wanted to basically test a private member function (in this case the timeout) but couldn't. What I could do is wrap the public interface in a wrappwer and then just pass the wrapper a default value. This allows me to be a little bit more reobust and allows me to be able to experiment with the timeout in another public interface. The caller does not know that this is happeing under the hood, but I know this deligation is happening and that allows me to be able to compse a function that does exactly what I want and at the same time allows me to test my code a little bit better.

Overall a cool technique I'll leverage more consciously. I've used the proxy/adapter pattern before to solve problems like this, but I think I sumbled on it more than I did it diliberatley.