Go Training Reflection
So it looks like this module is going to work with recursion and the testing of the reflection library. I like that. We defined a function walk() to the specification of the prompt walk(interface{}, func(input string). This allows us to work through generating the correct packages.

For now we are ignoring the interface entirely and just focusing on making sure our spy is in the right location at the right time. This allows us to grow our function call over time in a test driven manner.

Note to self: Again, don't start writing the solution first! Simplest possible solution solution to make the tests pass!