Go Training Reflection
I don't fully understand what is happening here. The reflect.valueOf() makes sense. It must return whatever the interface is. Each interface must have a function method called Field() that returns data for a field in the struct based on its positional area. Then we call .String() on that to cast it to what we want, or probably fail if there is no .String() method present.

So its allowing me to create a field that allows for the data to be built as we expect