Go Training Reflect
Wow. A lot of code here to make this work, but worth it. So we are defining a set of test cases and we are leveraging those tests (and sub tests) in the test base. We are passing in a struct that we defined in the test case with what we are expecting from our function call. We have not changed the underlying code to do anything different. Our function just basically gathers the the string. Nothing fancy here. When the walk() function is called it will just call that inline method that we defined in the tests to append to our array and dowhat we expect.