Go Train Reflection
Another win for the reflect library I guess. This time I was trying to create a new struct field that contained a number instead of all strings. This meant that I could now have a case wehre I'm not just checking string values. My code did not account for this. The field.Kind() will allow for me to look up the data type of the reflected object.

In short
* if you want to a type of reflect use reflect.ValueOf()
* if you want the data type of the reflected object you need to call the .Kind() method
* if you want the value of a field of the reflected object you call .Field() with the positional arguement of the field you want