Go Train Reflection
So two new keywords again  for reflect. I can find if my data is a struct type by just asking reflect if the type if reflect.Struct. This allows me to do recursion and thankfully my recursive routine isn't crazy so it should return as I expect. I had first tied to see if the field had its own .NumFields() which is the wrong way and non idomatic. What I should have done was what was in this commit and just ask the reflect library to tell me what the type was. Now we have a recursive method that is ready for action.