Go Training Reflection
I ALMOST had the right answer without reading further into the book. I did have the if statement. I did have the val.kind(). I was under the assumption that I had a reflect.Pointer that I could use... it was reflect.Ptr(). :angry_face: But still cool.

Since we have an interface passed into this function we need to understand what the interface is going. Walk does not reference that a pointer can be passed but go allows for interfaces and pointers to be used interchangeably which can lead to confustion. In our test we define the memory address of a struct and pass that to the walk() function. This means that we need to see what was passed in and if that was a pointer. If that happened to be a pointer then we call reflect's interface and pull back the element that the pointer is referencing. This is dereferencing without the * keyword which is kind of nice if you are doing type related lookups.