Go Train Reflection
Okay cool so we can return reflect.Value to get our underlying type that we need to work with. This means that the type we are currenly needing will allow us to return the reflected value back from a fuction. This would be great in understanding how to get back the type of an object if that object should arise.