Go Train Reflection
Sometimes the inelegant solutions are the best. Instead of wanting to try to find out the new type that is present (slices in this case) is usable we instead just create an entirely new code path for it. It can create some terse code but at the same time there is some elegance in the layout. There has to be a way though that we can iterate over the slice and drop down to the main field set.