Go Training Reflection
Another issue see when I'm trying to do some work is forgetting some of the patterns that I have learned in the past. I like my switch statements to only set values or become statemachines but I often time do not do real work in them because it can lead to confusion and the having of a fuction do more than one thing. In the case of a state machine this is fine because you are delegating behavior but I don't tend to pull the pattern unless the state machine is the obvious answer.

In this code pulling out that type so we check it first and then do work in each case is an example of a recursive machine that is not really a statemachine. Each frame will be put on the stack to do work until it reaches the string base case. Once that happens then everything will return and the loopwill complete. It keeps the code clean as well to walk all of the elements of the struct but at the same time it makes things a little messy in terms of branching code paths.