Go Training Reflection
Okay so we got to where I was trying to go but did it in more steps. Which is fine. I should have done that anyway. I love TDD but I still just want to code a solution instead of making the tests pass first and then refactoring. Its something I need to work on but I know I'm light years ahead of everyone else on it. Trial and error are all part of experimentation so I'm glad that we can take this time to do this kind of practice and really call out these kinds of problems.

Go is really cool because now I can store in a variable the number that I'm using in my loop iteration. I was going to do this, but not as cleanly early in the code. The one that caught me off guard was the fuction that is really a variable that will return the reflection unit that I can use in my loop. It may be that this is early morning, but I didn't see that pattern right away. I'm glad that we have it because we now moved the processing logic out of the loop and into the body of the function.