Go Train Reflection
Did it again! Tried to go right into the solution. I started easy enough to do the detection of the map via keys but then just jumped right into the general solution. I should have put the for loop in place and done the right thing. But for this we can see that the kind allows us to get the MapKeys() and then use that to look up the key and then convert that to an interface that we pass recursively to walk and move from there.