Go Training Reflection
Okay taking a step back from our beautiful solution to make it terse again. I like the abstraction of our walk function to a new function that really just encapsulates the walk function. It makes it a bit more obvious that we are doing some kind of work there. Due to how the map has keys instead of positional index we need to now put everything to the previous method and move back from there.