Go Training Reflection
Added in the reflected information as I expected. We needed to pull the map out into its own test case because keys in a map are not expected to return in the same order. So we just used the needle in a haystack pattern to look in our array for the key that we request and return that to the caller ie the test.