Go Training Reflection
I hate the go channel syntax. Utterly confusing

```
for v, ok := val.Recv(); ok; v, ok = val.Recv() {
    walk(v.Interface(), fun)
}
```
So what I'm seeing here is that we are just using val.Recv() to use reflection to read from the channel. Then we check if that chanel is good. Then we call Recv() again. So basically this is just an array with ok in the middle to make sure all of our functions behave correctly. There has to be better syntax as we don't always know how many times that our channel will be written to.