Go Training Reflect
Go really is a lot like Javascript and Ruby. You can use anon Call operators to call functions. The only caveat is that you don't know the results ahead of time like you would if you were calling the function normallly. Reflect simply does not have that kind of information at the ready. So instead it can return an array of results that the function may have returned by using the val.Call(nil) (or data if you have it) interface and returning the values returned. This would be something to play with to get a better understanding of how it works internally but for now I think this is a good primer.