Go Train Sync
Okay now we are getting into the meat of things. We wrote a test to allow for a parallel execution of the counter. Sadly we couldn't make that work as expected as there was no lock in place. We had to do two things: Update our test to wait for things to complete and update our code to make sure we are locked.

In order to make sure that we waited correctly we used golang's sync.WaitGroup() package. This allows us to create a wait group, set the amout of iterations we wanted to wait for in this case 1000 and then we send that as a pointer into the go function. When we call counter.Inc() we also send a signal to the waitGroup that the work has been finished with w.Done(). After the for loop we have our wg.Wait() waiting for all of our iterations to complete. This is the same as having a collector or master node waiting for all parallel execution to complete.

In the real code space we make sure that we lock everything down. Since counter is essentially global we have to make sure that no two treads can make changes to a single piece of data. We do that through the sync package's mutex. We add it to the data structure and because it is global it makes sure that we are not doing anything tricky. When we call Inc() we make sure to invoke the lock and call defer to unlock right after as we want the unlock to happen after the function completes. This protects us from multiple threads writing to the same data and stops crazy things from happening.

The tests now pass.