Go Training Sync
Another tool for the tool chain. ```go vet```. This will check for silly things in the code and help us to deterine if we can fix them. In this case it would be copying a mutex from one struct to another.

In order to get around this we need to do something a little more intelligent. Since we are dealing with a mutex in a struct we want to deal with the **pointer** value only and not the standard struct we are using now. By passing the memory address of the counter to the call to assertCount we are able to leverage the pointer and make things correct.

Go has several of these lurking errors in the code. This is the same as with C++. I am thankful of tools like go vet that can stop you from doing stupid things. Now if only I can take the time and remember to use the tooling.