Go Training Sync
Finally using a factory method. In the previous code we were passing the memory address to the assert count method. This is all well and good but we can do better. And by better I mean we can create a factory that will create the pointer for us. This makes better sense. I will like to see how we can leverage this to build any types. Right now we are not passing in build arguments so I will be interested to see how that gets handled. It could be with a default interface and an anon struct that contains the vaules. This will allow us to just read the struct for values and return the new object as expected.