Go Training Context
We are now talking about context within a server (or any process could be a go routine) in golang. We created a simple server and mocked out the core data store. This store could be a database for all we care. We have our server get called using httpTest.NewRequest and httpTest.NewRecorder() to mock the request and response correctly. Then we send that to our server to have it make the call.

Once the call is made we look at the response body and compare it with what was supposed to be in the server. In this case it was our stubstore in the test. Using the interface instead of a concrete type allowed us to mock out a ton of behavior and that is a good thing.