Go Train Context
We have added a new test for timing out a read by using go's context construct. The Context lives on the request and is helpful to have the server make decisions. If for whatever reason our context timesout we can look at our select and call store.Cancel(). This is helpful as it allows us to make the right steps forward when dealing with the context of a timeout.

We are now leveraging goroutines here. We are getting data to the server. We are reading that data through a channel. The channel will listen in the go routine for data from store.Fetch(). This data will then be piped to the select channel to pull in the data that we expect. We then send it to the print writer and then back to the client. This is helpful for when we need to stop long running processes on the server by imposing a time out. We can pull out the request's context and have that context be closed without disturbing any of the main data processing.

Spring handles this for us most of the time, it is kind of nice (and scary) that go allows us to manage this ourselves.